requirements.txt
requests>=2.31.0
sseclient-py>=1.7.2
flask>=3.0.0
flask-socketio>=5.3.6 